Move aura-api response data classes
